* Refactor use cases to accommodate unit tests * Implement unit tests for use cases * Implement ErrorMsgConstants

+++ app/src/main/java/dev/arch3rtemp/contactexchange/domain/usecase/DeleteCardUseCase.java
@@ -1,5 +1,7 @@
 package dev.arch3rtemp.contactexchange.domain.usecase;
 
+import static dev.arch3rtemp.contactexchange.domain.util.ErrorMsgConstants.MSG_ID_MUST_BE_POSITIVE;
+
 import dev.arch3rtemp.contactexchange.domain.repository.CardRepository;
 
 import javax.inject.Inject;
@@ -16,6 +18,7 @@ public final class DeleteCardUseCase {
     }
 
     public Completable invoke(int id) {
+        if (id <= 0) return Completable.error(new IllegalArgumentException(MSG_ID_MUST_BE_POSITIVE));
         return repository.deleteCard(id);
     }
 }
